+++ src/xrdtiffoperations/imagemodel/ifd/fields/FieldTypes.java
@@ -4,11 +4,11 @@ public class FieldTypes {
 
     /////////// Constants ///////////////////////////////////////////////////////////////////
 
-    public static int EIGHT_BIT_UNSIGNED_INT = 1;
-    public static int ASCII = 2;
-    public static int SIXTEEN_BIT_UNSIGNED_INT = 3;
-    public static int THIRTY_TWO_BIT_UNSIGNED_INT = 4;
-    public static int RATIONAL = 5;
+    public static short EIGHT_BIT_UNSIGNED_INT = 1;
+    public static short ASCII = 2;
+    public static short SIXTEEN_BIT_UNSIGNED_INT = 3;
+    public static short THIRTY_TWO_BIT_UNSIGNED_INT = 4;
+    public static short RATIONAL = 5;
 
 }
